Handle immutability correctly when possible.
We actually know when you want to be immutable, which means we can
stop you from doing odd things like generating an accessor that
will try to write to that immutable field and we can stop you at
compile (read: accessor generation) time, which is nice.

We can also make sure to generate the correct private accessors
as well, no need to have an lvalue sub when the instance is
immutable.